refactor(cli): flatten cli/display into cli package
Move the stack/value formatting helpers from cli/display into cli
itself. With repl already collapsed in, display was the last nested
subpackage; consolidating leaves cli flat. The helpers no longer need
to be exported now that both callers live in the same package, so
Stack/Value become unexported printStack/formatValue.
